+++ src/yaplox/expr.py
@@ -11,6 +11,10 @@ class ExprVisitor(ABC):
     """This class is used as an Vistor for the Expr class"""
 
     @abstractmethod
+    def visit_assign_expr(self, expr: "Assign"):
+        raise NotImplementedError
+
+    @abstractmethod
     def visit_binary_expr(self, expr: "Binary"):
         raise NotImplementedError
 
@@ -26,11 +30,25 @@ class ExprVisitor(ABC):
     def visit_unary_expr(self, expr: "Unary"):
         raise NotImplementedError
 
+    @abstractmethod
+    def visit_variable_expr(self, expr: "Variable"):
+        raise NotImplementedError
+
 
 class Expr(ABC):
     @abstractmethod
     def accept(self, visitor: ExprVisitor):
         raise NotImplementedError
+
+
+class Assign(Expr):
+    def __init__(self, name: Token, value: Expr):
+        self.name = name
+        self.value = value
+
+    def accept(self, visitor: ExprVisitor):
+        """ Create a accept method that calls the visitor. """
+        return visitor.visit_assign_expr(self)
 
 
 class Binary(Expr):
@@ -70,3 +88,12 @@ class Unary(Expr):
     def accept(self, visitor: ExprVisitor):
         """ Create a accept method that calls the visitor. """
         return visitor.visit_unary_expr(self)
+
+
+class Variable(Expr):
+    def __init__(self, name: Token):
+        self.name = name
+
+    def accept(self, visitor: ExprVisitor):
+        """ Create a accept method that calls the visitor. """
+        return visitor.visit_variable_expr(self)