+++ src/yaplox/parser.py
@@ -1,7 +1,7 @@
-from typing import List
+from typing import List, Optional
 
-from yaplox.expr import Binary, Expr, Grouping, Literal, Unary
-from yaplox.stmt import Expression, Print, Stmt
+from yaplox.expr import Assign, Binary, Expr, Grouping, Literal, Unary, Variable
+from yaplox.stmt import Expression, Print, Stmt, Var
 from yaplox.token import Token
 from yaplox.token_type import TokenType
 
@@ -27,9 +27,30 @@ class Parser:
     def parse(self) -> List[Stmt]:
         statements = []
         while not self._is_at_end():
-            statements.append(self._statement())
+            if declaration := self._declaration():
+                statements.append(declaration)
 
         return statements
+
+    def _declaration(self) -> Optional[Stmt]:
+        try:
+            if self._match([TokenType.VAR]):
+                return self._var_declaration()
+            return self._statement()
+        except ParseError:
+            self._synchronize()
+            return None
+
+    def _var_declaration(self) -> Stmt:
+        name = self._consume(TokenType.IDENTIFIER, "Expect variable name.")
+
+        initializer = None
+
+        if self._match([TokenType.EQUAL]):
+            initializer = self._expression()
+
+        self._consume(TokenType.SEMICOLON, "Expect ';' after variable declaration.")
+        return Var(name=name, initializer=initializer)
 
     def _statement(self) -> Stmt:
         if self._match([TokenType.PRINT]):
@@ -48,7 +69,20 @@ class Parser:
         return Expression(expr)
 
     def _expression(self) -> Expr:
-        return self._equality()
+        return self._assignment()
+
+    def _assignment(self) -> Expr:
+        expr = self._equality()
+
+        if self._match([TokenType.EQUAL]):
+            equals = self._previous()
+            value = self._assignment()
+
+            if isinstance(expr, Variable):
+                name = expr.name
+                return Assign(name=name, value=value)
+            self._error(equals, "Invalid assignment target.")
+        return expr
 
     def _equality(self) -> Expr:
         expr = self._comparison()
@@ -118,6 +152,9 @@ class Parser:
         if self._match([TokenType.NUMBER, TokenType.STRING]):
             return Literal(self._previous().literal)
 
+        if self._match([TokenType.IDENTIFIER]):
+            return Variable(self._previous())
+
         if self._match([TokenType.LEFT_PAREN]):
             expr = self._expression()
             self._consume(TokenType.RIGHT_PAREN, "Expect ')' after expression.")
@@ -160,12 +197,10 @@ class Parser:
         self.on_token_error(token, message)
         raise ParseError(token, message)
 
-    def _synchronize(self):  # pragma: no cover
+    def _synchronize(self):
         """
         When the parser detects an error, try to recover and consume tokens until a
         ; or an statement is found.
-
-        This is not yet functional, and thus excludede in coverage tests.
         """
         self._advance()