+++ tests/test_interpreter.py
@@ -162,7 +162,12 @@ class TestInterpreter:
 
     @pytest.mark.parametrize(
         ("expression", "result"),
-        [("4 * 6 / 2", "12"), ("12 < 6", "False"), ("12 > 6", "True"), ("3+3", "6")],
+        [
+            ("4 * 6 / 2;", "12"),
+            ("12 < 6;", "False"),
+            ("12 > 6;", "True"),
+            ("3 + 3;", "6"),
+        ],
     )
     def test_interpret(self, mocker, expression, result):
         on_scanner_error_mock = mocker.MagicMock()
@@ -171,9 +176,9 @@ class TestInterpreter:
         scanner = Scanner(expression, on_error=on_scanner_error_mock)
         tokens = scanner.scan_tokens()
         parser = Parser(tokens, on_token_error=on_parser_error_mock)
-        expr = parser.parse()
+        statements = parser.parse()
 
-        result = Interpreter().interpret(expr)
+        result = Interpreter().interpret(statements)
 
         assert result == result
 
@@ -182,17 +187,44 @@ class TestInterpreter:
         on_parser_error_mock = mocker.MagicMock()
         on_interpret_error_mock = mocker.MagicMock()
 
-        expression = '0 + "Foo"'
+        expression = '0 + "Foo";'
 
         scanner = Scanner(expression, on_error=on_scanner_error_mock)
         tokens = scanner.scan_tokens()
         parser = Parser(tokens, on_token_error=on_parser_error_mock)
-        expr = parser.parse()
+        statements = parser.parse()
 
-        Interpreter().interpret(expr, on_error=on_interpret_error_mock)
+        Interpreter().interpret(statements, on_error=on_interpret_error_mock)
 
         # There will be an error
         assert on_interpret_error_mock.called
         assert "Operands must be two numbers or two strings" in str(
             on_interpret_error_mock.call_args
         )
+
+    def test_assignment(self, mocker):
+        on_scanner_error_mock = mocker.MagicMock()
+        on_parser_error_mock = mocker.MagicMock()
+        on_interpret_error_mock = mocker.MagicMock()
+
+        lines = [
+            "var a = 0;",
+            "var c = a;",
+            "var b;",
+            "a = 3 + 6;",
+            "b = 3 / 6;",
+            "a = a + b;",
+            "print(a);",
+            "a;",
+        ]
+
+        expression = "\n".join(lines)
+
+        scanner = Scanner(expression, on_error=on_scanner_error_mock)
+        tokens = scanner.scan_tokens()
+        parser = Parser(tokens, on_token_error=on_parser_error_mock)
+        statements = parser.parse()
+
+        result = Interpreter().interpret(statements, on_error=on_interpret_error_mock)
+
+        assert result == 9.5