+++ CHANGELOG.md
@@ -10,10 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Completing chapter 8
-- Completed upto chapter 8.1.3
+- Completed upto chapter 8.4.2
 - Added Stmt class
 - Add isort to ast generator
 - Make imports configurable in ast generator
+- Create Environment
+- Variables work!
 
 ## [0.0.4] - 2020-08-23